start work on feature organisation. i think this will work out. for now, time to sleep

+++ src/main/scala/de/sciss/mutagentx/visual/impl/VisualConstantImpl.scala
@@ -27,7 +27,7 @@ object VisualConstantImpl {
 
     private def mkName(): Unit = _name = {
       val mathContext = new MathContext(6, RoundingMode.DOWN)
-      val bigDecimal = BigDecimal(_value, mathContext)
+      val bigDecimal = BigDecimal(_value.toDouble, mathContext)
       bigDecimal.bigDecimal.toPlainString
     }
